+++ ChangeLog.md
@@ -1,5 +1,11 @@
 # Changelog for epi-sim
 
+## 0.1.3.0
+
+- Implement a birth-death-sampling-occurrence model in `Epidemic.BirthDeathSamplingOccurrence`.
+- Implement a visualisation of the birth-death-sampling LTT in `vis/birthDeathSamplingVisualisation.R`.
+- Implement a birth-death-sampling model in `Epidemic.BirthDeathSampling`.
+
 ## 0.1.2.0
 
 - Implement a visualisation of the infection network in `vis/infection-tree.vg.json`.